+++ richat/src/bin/richat.rs
@@ -7,7 +7,7 @@ use {
     },
     richat::{
         channel::Messages, config::Config, grpc::server::GrpcServer, pubsub::server::PubSubServer,
-        source::subscribe,
+        richat::server::RichatServer, source::subscribe,
     },
     richat_shared::shutdown::Shutdown,
     signal_hook::{consts::SIGINT, iterator::Signals},
@@ -55,6 +55,7 @@ fn main() -> anyhow::Result<()> {
     // Create channel runtime (receive messages from solana node / richat)
     let messages = Messages::new(
         config.channel.config,
+        config.apps.richat.is_some(),
         config.apps.grpc.is_some(),
         config.apps.pubsub.is_some(),
     );
@@ -108,6 +109,14 @@ fn main() -> anyhow::Result<()> {
         move || {
             let runtime = config.apps.tokio.build_runtime("richatApp")?;
             runtime.block_on(async move {
+                let richat_fut = if let Some(config) = config.apps.richat {
+                    RichatServer::spawn(config, messages.clone(), shutdown.clone())
+                        .await?
+                        .boxed()
+                } else {
+                    ready(Ok(())).boxed()
+                };
+
                 let grpc_fut = if let Some(config) = config.apps.grpc {
                     GrpcServer::spawn(config, messages.clone(), shutdown.clone())?.boxed()
                 } else {
@@ -129,7 +138,7 @@ fn main() -> anyhow::Result<()> {
                     ready(Ok(())).boxed()
                 };
 
-                try_join_all(vec![grpc_fut, pubsub_fut, metrics_fut])
+                try_join_all(vec![richat_fut, grpc_fut, pubsub_fut, metrics_fut])
                     .await
                     .map(|_| ())
             })